refactor: make DynamoDB the only idempotency key store
Removes the Redis side of the idempotency-key dual-write now that the
`idempotencyDynamoRead` flip has been verified in prod: post-flip, every
409 is preceded 1:1 by a DynamoDB conditional-check failure.

- checkIdempotencyKey / releaseIdempotencyKey: single awaited Dynamo call
- delete external/redis/actions/idempotencyKeys/
- move the storage-key format pin out of the misc-redis key test into the
  idempotency test, since it is no longer a misc-cache family
- rewrite the dual-write unit test for the single-store contract

Keeps the `idempotencyDynamoRead` edge config field: zod strips unknown
keys, so an admin save from this image would drop it from S3 and flip a
still-running blue instance back onto a Redis store nothing writes to.
