+++ server/src/internal/misc/idempotency/actions/releaseIdempotencyKey.ts
@@ -1,11 +1,9 @@
 import { releaseDynamoIdempotencyKey } from "@/external/aws/dynamodb/idempotencyKeys/operations/releaseDynamoIdempotencyKey.js";
-import { releaseRedisIdempotencyKey } from "@/external/redis/actions/idempotencyKeys/operations/releaseRedisIdempotencyKey.js";
 import type { AutumnContext } from "@/honoUtils/HonoEnv.js";
-import { isIdempotencyDynamoReadEnabled } from "@/internal/misc/miscellaneousEdgeConfig/miscellaneousEdgeConfigStore.js";
 import { buildIdempotencyStorageKey } from "../idempotencyKeyUtils.js";
 
-/** Mirrors checkIdempotencyKey: the authority store's delete is awaited, the
- *  other store's is fire-and-forget. Both swallow their own errors. */
+/** Frees the key so the client can retry. Swallows its own errors — a failed
+ *  release only costs the caller a retry window, never the request. */
 export const releaseIdempotencyKey = async ({
 	ctx,
 	idempotencyKey,
@@ -19,12 +17,5 @@ export const releaseIdempotencyKey = async ({
 		idempotencyKey,
 	});
 
-	if (isIdempotencyDynamoReadEnabled()) {
-		void releaseRedisIdempotencyKey({ storageKey });
-		await releaseDynamoIdempotencyKey({ storageKey });
-		return;
-	}
-
-	void releaseDynamoIdempotencyKey({ storageKey });
-	await releaseRedisIdempotencyKey({ storageKey });
+	await releaseDynamoIdempotencyKey({ storageKey });
 };
